Refactor all the common pagination code to base.js and add pagination for genes and variables list
