+++ static/js/base.js
@@ -153,7 +153,11 @@ require([
         sortList: [[5,1]]
     });
 
-    $('#cohort-table').tablesorter({
+    $('#cohort-table')
+        .on('sortStart', function() {
+          update_table_display();
+        })
+        .tablesorter({
         headers: {
             0: {sorter:false},
             7: {sorter: 'fullDate'}
@@ -168,6 +172,159 @@ require([
         },
         sortList: [[4,1]]
     });
+
+    // ------ pagination -------
+
+    if (document.readyState == 'complete') {
+        update_table_display();
+    } else {
+        document.onreadystatechange = function () {
+            if (document.readyState === "complete") {
+                update_table_display();
+            }
+        }
+    }
+
+    // change no of entries per page
+    $('.panel-body').on('change', '.items-per-page', function () {
+        items_per_page = parseInt($('.items-per-page :selected').val());
+        goto_table_page(1);
+    });
+
+    function update_table_display()
+    {
+        var total_items = item_list.length;
+        var total_pages = Math.ceil(total_items / items_per_page);
+
+        //change another part at here
+        if (total_items <= 0) {
+            $('.item-page-count').hide();
+            $('.no-item-page-count').show();
+            $('.paginate_button_space').hide();
+            $('.dataTables_length').addClass('disabled');
+            $('.sortable_table th').addClass('disabled');
+            $('.dataTables_goto_page').addClass('disabled');
+        }
+        else {
+            var page_list = pagination(page, total_pages);
+            var html_page_button = "";
+            for(var i in page_list){
+                if(page_list[i] === "..."){
+                    html_page_button += "<span class='\ellipsis\'>...</span>"
+                }
+                else{
+                    html_page_button += "<a class=\'dataTables_button paginate_button numeric_button"+ (page_list[i] == page ? " current\'":"\'") +">" + page_list[i] + "</a>";
+                }
+            }
+            $('.item-page-count').show();
+            $('.no-item-page-count').hide();
+            $('.paginate_button_space').show();
+            $('.dataTables_length').removeClass('disabled');
+            $('.sortable_table th').removeClass('disabled');
+            $('.dataTables_goto_page').removeClass('disabled');
+            $('.dataTables_goto_page .goto-page-number').attr('max', total_pages);
+            $('.total-item-count').html(total_items);
+            $('.paginate_button_space').html(html_page_button);
+        }
+
+        first_page_entry = ((page - 1) * items_per_page) + 1;
+        last_page_entry = Math.min(first_page_entry + items_per_page - 1, total_items);
+
+        if (total_items <= 0) {
+            first_page_entry = 0;
+            last_page_entry = 0;
+        }
+        else {
+            for (i = 1; i <= total_items; ++i) {
+                var item_index_id = "#item-index-" + i;
+                if (i >= first_page_entry && i <= last_page_entry) {
+                    $(item_index_id).show();
+                } else {
+                    $(item_index_id).hide();
+                }
+            }
+
+            $('.showing-page').text(first_page_entry + " to " + last_page_entry);
+        }
+
+        $('.prev-page').removeClass('disabled');
+        $('.next-page').removeClass('disabled');
+        if (parseInt(page) == 1) {
+            $('.prev-page').addClass('disabled');
+        }
+        if (parseInt(page) * items_per_page >= total_items) {
+            $('.next-page').addClass('disabled');
+        }
+    };
+
+    function goto_table_page(page_no){
+        page=page_no;
+        update_table_display();
+    }
+
+    $('.panel-body').on('click', '.goto-page-button', function () {
+        var page_no_input = $(this).siblings('.goto-page-number').val();
+        if (page_no_input == "")
+            return;
+        var page = parseInt(page_no_input);
+        var max_page_no = parseInt($(this).siblings('.goto-page-number').attr('max'));
+        if (page > 0 && page <= max_page_no) {
+            goto_table_page(page);
+            $(this).siblings('.goto-page-number').val("");
+        } else {
+            utils.showJsMessage("warning",
+                "Page number you have entered is invalid. Please enter a number between 1 and " + max_page_no, true);
+            $('#placeholder').hide();
+        }
+    });
+
+    $('.panel-body').on('click', '.paginate_button', function () {
+        var page_no;
+        if ($(this).hasClass('next-page')) {
+            page_no = parseInt(page) + 1;
+        } else if ($(this).hasClass('prev-page')) {
+            page_no = page_no == 1 ? 1 : page - 1;
+        } else if ($(this).hasClass('numeric_button')) {
+            if ($(this).hasClass('current'))
+                return;
+            page_no = $(this).text();
+        } else {
+            page_no = 1;
+        }
+        goto_table_page(page_no)
+    });
+
+    // Returns a list of page numbers to show pagination buttons
+    // c is current page, m is total pages
+    function pagination(c, m) {
+        var current = parseInt(c),
+            last = m,
+            delta = 2,
+            left = current - delta,
+            right = current + delta + 1,
+            range = [],
+            rangeWithDots = [],
+            l;
+        for (var i = 1; i <= last; i++) {
+            if (i == 1 || i == last || i >= left && i < right) {
+                range.push(i);
+            }
+        }
+        for(var i in range){
+            if (l) {
+                if (range[i] - l === 2) {
+                    rangeWithDots.push(l + 1);
+                } else if (range[i] - l !== 1) {
+                    rangeWithDots.push('...');
+                }
+            }
+            rangeWithDots.push(range[i]);
+            l = range[i];
+        }
+        return rangeWithDots;
+    }
+
+    // ------ end pagination -------
 
     $(document).ready(function(){
         if(sessionStorage.getItem("reloadMsg")) {
@@ -258,9 +415,6 @@ define(['jquery', 'utils'], function($, utils) {
         },
         removeCookie: function(name, path) {
             utils.removeCookie(name, path);
-        },
-        pagination: function(current_page, total_pages) {
-            return utils.pagination(current_page, total_pages);
         },
         gdcSchema: {
             "type": "array",
