+++ templates/cohorts/cohort_list.html
@@ -301,14 +301,11 @@
     </div>
 
     <script type="text/javascript">
-    var items_per_page=5;
-    var page=1;
     var cohort_list = {{ cohorts_listing|safe }};
+    var item_list = cohort_list;
     var shared_users = {{ shared_users|safe }};
     var from_workbook = {{ from_workbook|lower }};
     var previously_selected_cohort_ids = {{ previously_selected_cohort_ids|safe }};
-    var first_page_entry;
-    var last_page_entry;
     </script>
 {% endblock %}
