+++ templates/genes/genes_list.html
@@ -22,6 +22,7 @@
     <div class="panel penel-default">
         <div class="panel-body">
             {% if gene_fav_list %}
+                {% include "isb_cgc/pagination.html" %}
                 <table id="gene-list-table" class="table table-striped tablesorter">
                     <thead>
                         <tr>
@@ -33,7 +34,7 @@
                     </thead>
                     <tbody>
                         {% for genes_fav in gene_fav_list %}
-                            <tr>
+                            <tr id="item-index-{{ forloop.counter }}">
                                 <td class="control-col">
                                     <div class="dropdown">
                                         <a data-target="#" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false"><i class="fa fa-caret-down"></i></a>
@@ -62,6 +63,11 @@
         </div>
     </div>
 </div>
+
+<script type="text/javascript">
+var gene_list = {{ gene_fav_listing|safe }};
+var item_list = gene_list;
+</script>
 
 {% endblock %}
