+++ templates/workbooks/workbook_list.html
@@ -82,11 +82,8 @@
     </div>
 </div>
 <script type="text/javascript">
-var items_per_page=5;
-var page=1;
 var workbook_list = {{ workbook_listing|safe }};
-var first_page_entry=0;
-var last_page_entry=3;
+var item_list = workbook_list;
 </script>
 
 {% endblock %}
